# apps/web-deprecated/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Vox Prismatic",
        "version": "1.0.0",
        "description": "Content Creation Platform API"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local Development"
        }
    ],
    "security": [
        {
            "sanctum": []
        }
    ],
    "paths": {
...
        "/auth/login": {
            "post": {
                "operationId": "auth.login",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "remember": {
                                        "type": "boolean"
                                    }
                                },
                                "required": [
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "user": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "createdAt": {
                                                    "type": "string"
                                                },
                                                "isAdmin": {
                                                    "type": "boolean"
                                                },
                                                "stripeCustomerId": {
                                                    "type": "string"
                                                },
                                                "stripeSubscriptionId": {
                                                    "type": "string"
                                                },
                                                "subscriptionStatus": {
                                                    "type": "string"
                                                },
                                                "subscriptionPlan": {
                                                    "type": "string"
                                                },
                                                "subscriptionCurrentPeriodEnd": {
                                                    "type": "string"
                                                },
                                                "cancelAtPeriodEnd": {
                                                    "type": "boolean"
                                                },
                                                "trialEndsAt": {
                                                    "type": "string"
                                                },
                                                "trialNotes": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "id",
                                                "email",
                                                "name",
                                                "createdAt",
                                                "isAdmin",
                                                "stripeCustomerId",
                                                "stripeSubscriptionId",
                                                "subscriptionStatus",
                                                "subscriptionPlan",
                                                "subscriptionCurrentPeriodEnd",
                                                "cancelAtPeriodEnd",
                                                "trialEndsAt",
                                                "trialNotes"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "user"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Cookie-based session authentication using Laravel Sanctum via content_session cookie. For non-GET requests, fetch /sanctum/csrf-cookie and include X-XSRF-TOKEN header.",
                "in": "cookie",
                "name": "content_session"
            }
        },
...
        "responses": {
...
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            }
        }
    }
}